telegram-bot/requirements.txt
python-telegram-bot==20.7
httpx>=0.25.2,<0.26.0
python-dotenv==1.0.0